requirements.txt
fastapi>=0.109,<0.110
asyncpg>=0.29,<0.30
gunicorn>=22.0,<23.0
uvicorn[standard]>=0.27,<0.31
python-telegram-bot[job-queue]>=21.0,<22.0
pandas>=2.0,<3.0
joblib>=1.3,<2.0
pydantic>=1.10,<2.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<0.28
sentry-sdk[fastapi]>=2.8,<3.0
